fix: transfer_to retains Content-Type and respects existing info.
Relevant to #156
