+++ cloudvolume/__init__.py
@@ -4,7 +4,7 @@ from .provenance import DataLayerProvenance
 from .skeletonservice import PrecomputedSkeleton, SkeletonEncodeError, SkeletonDecodeError
 from .storage import Storage
 from .threaded_queue import ThreadedQueue
-from .txrx import EmptyVolumeException, EmptyRequestException, AlignmentError
+from .exceptions import EmptyVolumeException, EmptyRequestException, AlignmentError
 from .volumecutout import VolumeCutout
 
 from . import secrets
