+++ cloudvolume/exceptions.py
@@ -1,0 +1,18 @@
+
+
+class AlignmentError(Exception):
+  """Signals that an operation requiring chunk alignment was not aligned."""
+  pass 
+
+class EmptyVolumeException(Exception):
+  """Raised upon finding a missing chunk."""
+  pass
+
+class EmptyRequestException(Exception):
+  """
+  Requesting uploading or downloading 
+  a bounding box of less than one cubic voxel
+  is impossible.
+  """
+  pass
+
